Fix minor bugs of version 0.2.0.
