+++ lib/find.js
@@ -2,38 +2,14 @@
 //
 const path = require('path');
 const CSON = require('cson');
+const Util = require('./util');
 
 module.exports = class Find {
-  // if panel is <atom-panel class="tool-panel panel-bottm">, then return true.
-  static isAtomPanel(panel) {
-    // return (panel.tagName == 'atom-panel' && panel.className == 'tool-panel panel-bottom bottom')
-    return panel.className == "tool-panel panel-bottom";
-  }
-
-  // replace Tooltips's title with localized title
-  static setTooltipLabel(element, oldt, newt) {
-    if (!element) {
-      return;
-    };
-    const [tooltip] = atom.tooltips.findTooltips(element);
-    if (!tooltip) {
-      return;
-    }
-    if (tooltip.getTitle() == oldt) {
-      tooltip.options.title = newt;
-    } else {
-      let str = new String(tooltip.getTitle());
-      str.replace(oldt, newt);
-      tooltip.options.title = str.valueOf();
-    }
-    tooltip.fixTitle();
-    // TODO: ツールチップは内部で再表示されるため、パッケージ自身の修正が必要!
-  }
 
   // updateFindProject
   static updateFind(panel) {
     // if (!panel) { 1st }, if (panel) { 2nd }
-    if (panel && !this.isAtomPanel(panel)) {
+    if (panel && !Util.isAtomPanel(panel)) {
       return;
     };
     // "Find and Replace" panel
@@ -71,7 +47,7 @@ module.exports = class Find {
     // Line #1: Close Button
     const closeb = findPanel.getElementsByClassName("header-item close-button pull-right");
     if (closeb && closeb[0]) {
-      this.setTooltipLabel(closeb[0],
+      Util.setTooltipLabel(closeb[0],
         this.tmpltF.Find.optionTooltip.closePanel,
         this.defF.Find.optionTooltip.closePanel
         );
@@ -82,25 +58,25 @@ module.exports = class Find {
       // Tooltip title
       const btnRegex = buttons[0].childNodes[0];
       if (btnRegex) {
-        this.setTooltipLabel(btnRegex,
+        Util.setTooltipLabel(btnRegex,
           this.tmpltF.Find.optionTooltip.useRegex, this.defF.Find.optionTooltip.useRegex
           );
       };
       const btnCase = buttons[0].childNodes[1];
       if (btnCase) {
-        this.setTooltipLabel(btnCase,
+        Util.setTooltipLabel(btnCase,
           this.tmpltF.Find.optionTooltip.matchCase, this.defF.Find.optionTooltip.matchCase
           );
       };
       const btnOnly = buttons[0].childNodes[2];
       if (btnOnly) {
-        this.setTooltipLabel(btnOnly,
+        Util.setTooltipLabel(btnOnly,
           this.tmpltF.Find.optionTooltip.onlyIn, this.defF.Find.optionTooltip.onlyIn
           );
       };
       const btnWhole = buttons[0].childNodes[3];
       if (btnWhole) {
-        this.setTooltipLabel(btnWhole,
+        Util.setTooltipLabel(btnWhole,
           this.tmpltF.Find.optionTooltip.wholeWord, this.defF.Find.optionTooltip.wholeWord
           );
       }
@@ -110,11 +86,11 @@ module.exports = class Find {
     if (findNext) {
       findNext.textContent = this.defF.Find.button["find-next"];
       // Tooltip title
-      this.setTooltipLabel(findNext,
+      Util.setTooltipLabel(findNext,
         this.tmpltF.Find.buttonTooltip.findNext,
         this.defF.Find.buttonTooltip.findNext
         );
-      this.setTooltipLabel(findNext,
+      Util.setTooltipLabel(findNext,
         this.tmpltF.Find.buttonTooltip.findNextDisabled,
         this.defF.Find.buttonTooltip.findNextDisabled
         );
@@ -123,11 +99,11 @@ module.exports = class Find {
     if (findAll) {
       findAll.textContent = this.defF.Find.button["find-all"];
       // Tooltip title
-      this.setTooltipLabel(findAll,
+      Util.setTooltipLabel(findAll,
         this.tmpltF.Find.buttonTooltip.findAll,
         this.defF.Find.buttonTooltip.findAll
         );
-      this.setTooltipLabel(findAll,
+      Util.setTooltipLabel(findAll,
         this.tmpltF.Find.buttonTooltip.findAllDisabled,
         this.defF.Find.buttonTooltip.findAllDisabled
         );
@@ -142,11 +118,11 @@ module.exports = class Find {
     if (replaceNext) {
       replaceNext.textContent = this.defF.Find.button["replace-next"];
       // Tooltip title
-      this.setTooltipLabel(replaceNext,
+      Util.setTooltipLabel(replaceNext,
         this.tmpltF.Find.buttonTooltip.replaceNext,
         this.defF.Find.buttonTooltip.replaceNext
         );
-      this.setTooltipLabel(replaceNext,
+      Util.setTooltipLabel(replaceNext,
         this.tmpltF.Find.buttonTooltip.replaceNextDisabled,
         this.defF.Find.buttonTooltip.replaceNextDisabled
         );
@@ -155,11 +131,11 @@ module.exports = class Find {
     if (replaceAll) {
       replaceAll.textContent = this.defF.Find.button["replace-all"];
       // Tooltip title
-      this.setTooltipLabel(replaceAll,
+      Util.setTooltipLabel(replaceAll,
         this.tmpltF.Find.buttonTooltip.replaceAll,
         this.defF.Find.buttonTooltip.replaceAll
         );
-      this.setTooltipLabel(replaceAll,
+      Util.setTooltipLabel(replaceAll,
         this.tmpltF.Find.buttonTooltip.replaceAllDisabled,
         this.defF.Find.buttonTooltip.replaceAllDisabled
         );
