+++ lib/findProject.js
@@ -2,36 +2,14 @@
 //
 const path = require('path');
 const CSON = require('cson');
+const Util = require('./util');
 
 module.exports = class FindProject {
-  // if panel is <atom-panel class="tool-panel panel-bottm">, then return true.
-  static isAtomPanel(panel) {
-    // return (panel.tagName == 'atom-panel' && panel.className == 'tool-panel panel-bottom bottom');
-    return panel.className == "tool-panel panel-bottom";
-  }
-
-  // replace Tooltips's title with localized title
-  static setTooltipLabel(element, oldt, newt) {
-    if (!element) {
-      return;
-    }
-    const [tooltip] = atom.tooltips.findTooltips(element);
-    if (!tooltip) {
-      return;
-    }
-    if (tooltip.getTitle() == oldt) {
-      tooltip.options.title = newt;
-    } else {
-      let str = new String(tooltip.getTitle());
-      str.replace(oldt, newt);
-      tooltip.options.title = str.valueOf();
-    }
-  }
 
   // updateFindProject
   static updateFindProject(panel) {
     // if (!panel) { 1st }, if (panel) { 2nd }
-    if (panel && !this.isAtomPanel(panel)) {
+    if (panel && !Util.isAtomPanel(panel)) {
       return;
     }
     // "Find in Project" panel
@@ -69,7 +47,7 @@ module.exports = class FindProject {
     // Line #1: Close Button
     const closeb = findPanel.getElementsByClassName("header-item close-button pull-right");
     if (closeb && closeb[0]) {
-      this.setTooltipLabel(closeb[0],
+      Util.setTooltipLabel(closeb[0],
         this.tmpltP.FindProject.optionTooltip.closePanel,
         this.defP.FindProject.optionTooltip.closePanel
         );
@@ -80,21 +58,21 @@ module.exports = class FindProject {
       // Tooltip title
       const btnRegex = buttons[0].childNodes[0];
       if (btnRegex) {
-        this.setTooltipLabel(btnRegex,
+        Util.setTooltipLabel(btnRegex,
           this.tmpltP.FindProject.optionTooltip.useRegex,
           this.defP.FindProject.optionTooltip.useRegex
           );
       };
       const btnCase = buttons[0].childNodes[1];
       if (btnCase) {
-        this.setTooltipLabel(btnCase,
+        Util.setTooltipLabel(btnCase,
           this.tmpltP.FindProject.optionTooltip.matchCase,
           this.defP.FindProject.optionTooltip.matchCase
           );
       };
       const btnWhole = buttons[0].childNodes[2];
       if (btnWhole) {
-        this.setTooltipLabel(btnWhole,
+        Util.setTooltipLabel(btnWhole,
           this.tmpltP.FindProject.optionTooltip.wholeWord,
           this.defP.FindProject.optionTooltip.wholeWord
           );
@@ -105,7 +83,7 @@ module.exports = class FindProject {
     if (findAll) {
       findAll.textContent = this.defP.FindProject.button["find-all"];
       // Tooltip title
-      this.setTooltipLabel(findAll,
+      Util.setTooltipLabel(findAll,
         this.tmpltP.FindProject.buttonTooltip.findAll,
         this.defP.FindProject.buttonTooltip.findAll
         )
@@ -120,11 +98,11 @@ module.exports = class FindProject {
     if (replaceAll) {
       replaceAll.textContent = this.defP.FindProject.button["replace-all"];
       // Tooltip title
-      this.setTooltipLabel(replaceAll, 
+      Util.setTooltipLabel(replaceAll, 
         this.tmpltP.FindProject.buttonTooltip.replaceAll,
         this.defP.FindProject.buttonTooltip.replaceAll
         )
-      this.setTooltipLabel(replaceAll,
+      Util.setTooltipLabel(replaceAll,
         this.tmpltP.FindProject.buttonTooltip.replaceAllDisabled,
         this.defP.FindProject.buttonTooltip.replaceAllDisabled
         )
