+++ lib/main.js
@@ -25,13 +25,13 @@ class I18N {
     this.customMenuI18n = this.customMenuI18n.bind(this);
 
     // BUG when running spec, locale is not initialized
-    if (!atom.config.get('atom-i18n.customMenuI18nPath')) {
+    if (!atom.config.get('atom-i18n-beta.customMenuI18nPath')) {
       atom.config.set(
-        'atom-i18n.customMenuI18nPath',
+        'atom-i18n-beta.customMenuI18nPath',
         path.join(__dirname, '../def', 'custom_menu.cson'),
       );
     }
-    const LOCALE = atom.config.get('atom-i18n.locale');
+    const LOCALE = atom.config.get('atom-i18n-beta.locale');
     this.defM = CSON.load(
       path.join(__dirname, '../def', LOCALE, `menu_${process.platform}.cson`),
     );
@@ -53,9 +53,9 @@ class I18N {
     setTimeout(this.customMenuI18n, 3000);
     atom.commands.add(
       'atom-workspace',
-      'atom-i18n:open-custom-menu-i18n-file',
+      'atom-i18n-beta:open-custom-menu-i18n-file',
       () => {
-        atom.workspace.open(atom.config.get('atom-i18n.customMenuI18nPath'));
+        atom.workspace.open(atom.config.get('atom-i18n-beta.customMenuI18nPath'));
       },
     );
   }
@@ -77,12 +77,12 @@ class I18N {
     Keybind.localize(this.defK);
     TreeView.localize(this.defR);
     Notification.localize(this.defN);
-        // TODO localize more...
+    // TODO localize more...
     Util.handleConfigChange();
   }
 
   customMenuI18n() {
-    const customMenuCson = atom.config.get('atom-i18n.customMenuI18nPath');
+    const customMenuCson = atom.config.get('atom-i18n-beta.customMenuI18nPath');
     if (fs.existsSync(customMenuCson)) {
       this.customDefM = CSON.load(customMenuCson);
       Menu.localize(this.customDefM);
