+++ lib/util.js
@@ -37,4 +37,32 @@ module.exports = class Util {
     const option = configEnum.find((option) => option.value === locale);
     return option ? option.description : null;
   }
+
+  /*
+  * used in find.js & findProject.js
+  */
+  // if panel is <atom-panel class="tool-panel panel-bottm">, then return true.
+  static isAtomPanel(panel) {
+    // return (panel.tagName == 'atom-panel' && panel.className == 'tool-panel panel-bottom bottom')
+    return panel.className == "tool-panel panel-bottom";
+  }
+
+  // replace Tooltips's title with localized title
+  static setTooltipLabel(element, oldt, newt) {
+    if (!element) {
+      return;
+    };
+    const [tooltip] = atom.tooltips.findTooltips(element);
+    if (!tooltip) {
+      return;
+    }
+    if (tooltip.getTitle() == oldt) {
+      tooltip.options.title = newt;
+    } else if (tooltip.getTitle() !== '') {
+      let str = new String(tooltip.getTitle());
+      str.replace(oldt, newt);
+      tooltip.options.title = str.valueOf();
+    }
+    tooltip.fixTitle();
+  }
 }
